Apply all unapplied migrations, even in the past
Here is the problem. During development, you merge a migration which is
in your past. However, if you try to do `goose up`, the current goose
won't recognize the unapplied migrations, because it only goes up or
down the _current_ version. This may leave the database in inconsistent
state between developers or servers.

This change applies every unapplied version, even if it happens to be
in the past.